refactor(ui): move BetaTag to ui
