+++ client/src/menubar/menus/ShareMenu/ExportStreetmeter.tsx
@@ -1,6 +1,6 @@
+import { BetaTag } from '~/src/ui/BetaTag.js'
 import { Icon } from '~/src/ui/Icon.js'
 import { MenuItem } from '../MenuItem.js'
-import { BetaTag } from '../BetaTag.js'
 
 export function ExportStreetmeter() {
   return (
